Update read_msgf_data_from_DMS
The parameter data_package_num was not used, and the param DataPkgNumber did not exist in the function definition; changed the latter to the former. Modified param_file input checks so that the user does not have to provide a parameter file if there is only one. Use the PlexedPiper test data package in the example because it is smaller.